Confused fix of page breaks
So I think the problem was that if a page break was in a paragraph,
theres another rule that will end the paragraph and then start a new
one after the page break. But, it would not work if there wasn’t
content, so I added a non breaking space to force it to have content.
I’m still confused though.