trying out separating filetypes by internal behavior
I have this idea that if they all respond to the same methods
in terms of input and response, that then they can be used interchangeably
once they are setup. Calling .transform on a TEI file and a CSV file
would look identical from the outside but then they will run different
scripts / actions behind the scenes to accomplish this.
We shall see how it goes!