changes tei_to_es to return multivalued
this means that if you need subdocs generated, there is now at
least a rudimentary way to put them in, but the system is really
not very well thought out. mostly a rush, proof-of-concept type of
job so that we can accommodate it

problems on the horizon: what if project has different types of docs
which require completely different types of sub-docs? can work with
this system, probably, but it will mean lots of conditionals
and many lines of code

another problem: the general fields can only be used for the primary
doc, not the subdocs. Is this a problem?