adds more api fields to xpaths and file type methods
this looks like a lot of changed code but for the most part,
it is actually just adding more default methods for all formats,
as well as xpaths (even if blank)

this means that now, for your average field, if a data repo
adds an appropriate xpath, you may reasonable assume that the ES
JSON will be generated with it, whereas before it may not have shown
up automatically if that particular field was not commonly used. This
was leading to confusion about how and when to override different aspects,
which this commit attends to standardize.

Additionally, while in here, I took the liberty of switching get_text
and get_list to using keyword arguments. This will be a breaking change
for those projects which have called those methods (probably anything
with a personography file / subsection).

Also used this time to add to a few fields which had previously not been
implemented, like date_not_before and date_not_after