modify item payload preparation to handle already prepared values
All values are now explicitly formatted with property_id etc. as they are updated, a necessary change to allow for updating fields that were previously missing. The method (from the omeka_s_tools module) to check new item payloads against the schema needed to be updated to handle values that are already formatted